docs: publish doxygen site with github pages

+++ include/CosmicEngine/models/font/font.cpp
@@ -1,3 +1,8 @@
+/**
+ * @file font.cpp
+ * @brief Implements the font resource used by the engine text renderer.
+ */
+
 #include "font.hpp"
 
 #include <glad/glad.h>
@@ -66,7 +71,7 @@ namespace CosmicEngine
                 texture,
                 glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
                 glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-                face->glyph->advance.x
+                static_cast<unsigned int>(face->glyph->advance.x)
             };
 
             characters_resources.insert(std::pair<char, TextCharacter>(c, character));
